+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## [1.0.2] - 19 October 2022
+
+-   Slide Drawer ( LTR , RTL ) Direction Support
+
 ## [1.0.1] - 23 August 2022
 
-- added dynamically change drawer background image 
+-   added dynamically change drawer background image
 
 ## [1.0.0] - 9 July 2021
 
-- The first flutter slider drawer plugin released.
+-   The first flutter slider drawer plugin released.